Add files via upload
Based on the results of our work,  the following data-driven recommendations can be presented. Let's start with recommendations for regional markets. Mexico represents a promising market with a willingness to pay, as it has the highest average purchase volumes and purchasing power. Although the company has a low market penetration in the country, it is still a very promising market. However, these are preliminary conclusions that need to be verified with a larger sample.

Recommendations for distribution channels. Analytics show significant interest in the website, as evidenced by the large number of visits. However, this does not translate into sales, as the average number of purchases through the website is not very high.

Next, recommendations on products. The highest demand is for categories such as wine and meat products. Therefore, the company might consider expanding its assortment in these categories and offering more premium products to further increase sales. Additionally, some high-income customers (over 100,000 $) may be interested in such products.

Recommendations for marketing campaigns. In future campaigns, it might be worth trying to replicate the strategy of the last campaign, as it proved to be the most effective. And since we did not find strong connections between the geographical region and the campaign's success, future campaigns could be tested in smaller regions first, and if successful, rolled out across all regions. This will help save on unsuccessful campaigns and increase flexibility in campaign planning.

Recommendations on customer loyalty. It is advisable to introduce a loyalty or rewards program for participants to increase the popularity of the supermarkets and the frequency of purchases. This is because engagement among participants is quite low, as many have been members for over a year, but the average time since their last purchase is about 50 days. Given that the store offers a wide range of fresh products, such as fruits, fish, and meat, it is assumed that these products are purchased if not daily, then at least weekly or every two weeks.